fix toFacility() so it preserves numbers
Fix a bug in toFacility that resulted in toFacility(0) returning the
string 'LOG_KERN' (a reverse lookup). Because toFacility is used
by syslog.open() this was resulting in all numeric (aka, pre-resolved)
log facilities being mapped to a string which always converted to 0.

This should fix the problem reported in #8 and first discovered in
baudehlo/Haraka#1145